+++ includes/minishell.h
@@ -56,6 +56,8 @@ void		debug_2d_array(char **array);
 void		debug_token_dq(t_deque *deque, const char *str);
 void		debug_token_dq_node(t_deque_node *node);
 void		debug_print_ast_tree(t_ast *root, const char *str);
+char		*get_tree_symbol(bool is_rhs);
+char		get_tree_space(bool is_rhs);
 
 /* destroy */
 void		destroy_context(t_context context);
