+++ srcs/parse/abstruct_syntax_tree.c
@@ -36,28 +36,6 @@ arithmetic
 //	ft_dprintf(2, "token_str:%s\n", str);
 //}
 
-t_node_kind	convert_kind_token_to_node(t_deque_node *token_node)
-{
-	t_token			*token;
-	t_token_kind	token_kind;
-
-	if (!token_node)
-		return (0);//todo: tmp
-	token = token_node->content;
-	token_kind = token->kind;
-	if (token_kind == TOKEN_KIND_OP_PIPE)
-		return (NODE_KIND_OP_PIPE);
-	if (token_kind == TOKEN_KIND_OP_OR)
-		return (NODE_KIND_OP_OR);
-	if (token_kind == TOKEN_KIND_OP_AND)
-		return (NODE_KIND_OP_AND);
-	if (token_kind == TOKEN_KIND_PAREN_LEFT || token_kind == TOKEN_KIND_PAREN_RIGHT)
-		return (NODE_KIND_SUBSHELL);
-	if (token_kind == TOKEN_KIND_WORD)
-		return (NODE_KIND_COMMAND);
-	return (NODE_KIND_NONE); //todo: tmp
-}
-
 // <op_list>
 // expr	= <com_list> '&&' <com_list>
 //		| <com_list> '||' <com_list>
@@ -105,49 +83,6 @@ t_ast	*create_command_list_node(t_deque_node **token_node)
 	return (left_node);
 }
 
-static t_token	*dup_token(t_token *token)
-{
-	t_token	*new_token;
-
-	new_token = init_token_struct();
-	new_token->str = x_ft_strdup(token->str);
-	new_token->kind = token->kind;
-	new_token->quote = token->quote;
-	new_token->concat_next = token->concat_next;
-	return (new_token);
-}
-
-static t_deque_node	*deque_dup_node(t_token *token)
-{
-	t_deque_node	*new_node;
-	t_token			*new_token;
-
-	new_token = dup_token(token);
-	new_node = deque_node_new((void *)new_token);
-//	ft_dprintf(2, "dup_cmd:%s\n", new_token->str);
-	if (!new_node)
-		ft_abort();
-	return (new_node);
-}
-
-static void	dup_command_from_tokens(t_deque *command, t_deque_node **token_node)
-{
-	t_deque_node	*dup_node;
-	t_node_kind		node_kind;
-
-	if (!token_node)
-		return ;
-	while (*token_node)
-	{
-		node_kind = convert_kind_token_to_node(*token_node);
-		if (node_kind != NODE_KIND_COMMAND && node_kind != NODE_KIND_NONE)
-			return ;
-		dup_node = deque_dup_node((t_token *)(*token_node)->content);
-		deque_add_back(command, dup_node);
-		*token_node = (*token_node)->next;
-	}
-}
-
 // [echo] [a] [b] [|] [cat] [Makefile]
 
 // <com or subshell>
@@ -163,7 +98,7 @@ t_ast	*create_command_or_subshell_node(t_deque_node **token_node)
 
 	if (!*token_node)
 	{
-		ft_dprintf(STDERR_FILENO, "%s\n", 
+		ft_dprintf(STDERR_FILENO, "%s\n", \
 				"parse NULL: syntax error near unexpected token");
 		return (NULL);
 	}
